Don't throw a ProBError if a command causes warnings (but no errors)
This is an initial fix for PROB2UI-345, so that machines that cause
warnings can be opened at all. Now if a command produces only warnings
(and no errors), the warnings will only be logged. In the future there
should be a way to programmatically query these warnings, so that they
can be displayed properly in the UI.