Deprecate GetMachineStructureCommand and related classes
There are already other APIs that serve similar purposes, and they are
generally better than GetMachineStructureCommand.

This API was originally added to support the state view in the
ProB 2 UI. This is no longer necessary - the UI now uses BVisual2 to
get the machine structure (which is the same API used by the Tcl/Tk and
Rodin state views).