Deprecate ModelChecker and related methods
The ModelChecker class offers almost no additional features over the
individual checking job classes, and it makes it impossible for the
calling code to control the executor that is used, or for multiple
checkers to share an executor. ModelChecker also provides no way to
shut down the automatically created executor, which causes the JVM to
never terminate once a ModelChecker instance has been created.