Migrate from java to java-library plugin
This allows explicitly declaring dependencies as either api (for use by
both ProB 2 and its users) or as implementation (for use only by ProB 2
internally, and not for users of ProB 2).

For now, all (non-test) dependencies are declared as api, to avoid
breaking existing code that uses ProB 2. Some of these dependencies
will be changed to implementation in the future, if they are not used
in any of ProB 2's external APIs.