Remove serialization support for CSP formulas
CSP formulas were serialized using Gson with default settings, which is
highly dependent on the internal structure of the class in question.

The fields of the CSP class have changed many times since the
serialization implementation was added, meaning that older serialized
CSP formulas cannot be deserialized in newer versions. One of the
fields (cspmfPath) also contains a local file path, meaning that CSP
formulas can only be deserialized on the machine that serialized them.

These are fairly major issues, so it's safer to remove CSP
de/serialization support entirely for now. If someone needs it, it
would be better to reimplement it in a more reliable way.