Remove ExploreStateCommand log message about no constants being found
The log is normally not seen by the user, so it is not a good place to
report a possible error like this to the user. Some users of the API
might not care about this error at all. It's better to leave it up to
the actual frontend/UI (ProB 2 UI, Jupyter kernel, etc.) to report this
kind of error.