Allow listening for changes to the state space (again)
A similar mechanism was already implemented in the past, but has been
removed in 99f86eed04087235a59de7416173fa4e99846dda.

This allows detecting when new states and/or transitions have been
added to the state space, which can for example be used by the
ProB 2 UI to detect when to update the statistics view (which shows the
total number of states and transitions in the state space).