Fix handling of special ops in getTransitionsBasedOnParameterValues
When called for $setup_constants or $initialise_machine, the previous
implementation silently ignored the parameter values and executed the
operation with arbitrary constant/variable values. This is probably not
the expected behavior in most cases.

Now the method throws an error if it is called with a special operation
name and asks the user to call transitionFromPredicate instead, which
works correctly with special operations.