Move ProB 2 CLI into a separate subproject from main library
This way the classes and dependencies that are only needed by the CLI
are not included in the library.

The old de.prob.Main class is kept for compatibility, because it's
often used by ProB 2 users to get version information about ProB 2 or
the location of the ProB home directory.

At the moment de.prob.Main also holds some other state, like a global
injector and the maxCacheSize setting. This should probably be
refactored.

Closes PROBCORE-968.