Move setting of prob.stdlib from CommandLineMain to ClassicalBFactory
This way the setting is closer to where it's actually used. It also
means that other code that uses ProB 2 no longer needs to set
prob.stdlib manually. This was previously necessary because prob.stdlib
was only set in the ProB 2 CLI code, which isn't executed when ProB 2
is used as a library. (The new implementation only sets prob.stdlib if
it isn't already set, so code that intentionally sets a different
prob.stdlib will continue to work correctly.)