Separate injectors for command line parsing and actual ProB execution
This will allow configuring the ProB injector based on command line
options. This would not be possible with a single injector, because the
injector has to be created already to parse the command line, and it's
not possible (at least not very well) to change the configuration of an
existing injector.