Get JsonManager Gson instance from Context instead of via injection
This avoids relying on a single central Gson configuration and instead
allows different file types to use differently configured Gson
instances. The metadata is also written using an internal Gson instance
that is separate from the one used for the main data, to ensure that
the type adapters needed for the metadata are always present.