Optimize multi-thread use of AnimatorImpl by narrowing synchronized
The lock is now held only during the actual communication with probcli.
Result and error handling is now done after the lock is released, which
is safe, because it doesn't require communicating with probcli.

This improves performance a bit if multiple threads use an animator at
the same time, especially if some of the commands take a long time to
process their results.