Deprecate GetErrorsCommand in favor of GetErrorItemsCommand
Also removing the corresponding unit test. I don't think it makes much
sense to write a new test for GetErrorItemsCommand - the command is
already tested sufficiently as part of any test that communicates with
the CLI. If anything, a separate unit test will become outdated if the
CLI's error format changes.