Fix dot process not always being destroyed if thread is interrupted
When a thread is interrupted while waiting on blocking IO, an
InterruptedIOException is supposed to be thrown, but apparently this
doesn't work reliably on many platforms - see:
https://stackoverflow.com/a/4426050

As a workaround, move the stdout reading into a background thread, and
have the thread that runs DotCall wait on that thread. This way the
DotCall thread can still reliably receive interrupts and destroy the
dot process when interrupted, which will cause the other IO threads to
fail and terminate.