Add initial Jackson-based replacement for JsonManager
The new JacksonManager is mostly similar to the old JsonManager. Some
practically unused parts of the JsonManager API are no longer present
in JacksonManager. JacksonManager also has less special behavior around
the metadata field - the metadata is no longer separated from the
object when reading and added back when writing. Instead, it's expected
that the root object also has a real metadata field in Java.