Remove broken leftover code for ignoring interrupts while receiving
The ProBConnection.busy variable is no longer updated since we replaced
our custom reading code with Scanner. This caused all interrupts to be
ignored all the time, not just while reading an answer. It's not clear
why this special case was needed at all, and things seem to work fine
without the busy check, so remove it for now.