Remove no longer used ModelTranslationError throws declarations
Everything that previously threw ModelTranslationError now throws
ProBError instead (which is unchecked), so the ModelTranslationError
throws declarations can be removed.

This change may break compilation for code that explicitly catches
ModelTranslationError, because the compiler will complain that the
caught exception can never be thrown. This can be fixed by changing the
catch blocks to ProBError instead (or removing the
ModelTranslationError catch, if appropriate).