Change CompareTwoOperations to use new simpler 3-arg Prolog command
Compared to the old 5-argument version, this one no longer takes a list
of old identifiers and no longer returns an old/new identifier mapping.
The Java side now takes care of building this mapping.

The main motivation behind this change is to avoid the JSON string
construction for the Delta output argument. The mapping could be
returned as a Prolog list of key-value pairs instead, but because the
mapping can also be constructed on the Java side, it's easier to do
that instead of passing it through Prolog unnecessarily.