Un-deprecate some formula constructors and related methods
Sometimes one really doesn't care about the formula's expansion mode,
for example when the formula is never actually evaluated. In those cases
it's pointless to force the caller to set an expansion mode that will
never actually be used.

Methods like State.eval(String) remain deprecated, because they always
evaluate the formula, so the expansion mode is always relevant.