Fix confusing naming of model checking option FIND_OTHER_ERRORS
The new option IGNORE_OTHER_ERRORS has exactly the same meaning as the
old FIND_OTHER_ERRORS, but with the correct name and description. The
old, confusing name is deprecated, but stays for now.

Most code should not be affected by this, because the checkOtherErrors
method was already correctly named.

Fixes hhu-stups/prob-issues#160.