Deprecate evaluating overloads of transition list getters
These APIs are somewhat confusing. They overload a simple no-arg getter
with additional code that may sometimes do extra work, making it unclear
if the methods are cheap to call or not. The code flow of the no-arg
getter was also less obvious as a result. Additionally, the second
argument was ignored if the first argument was false.

For callers, the extra arguments are easy to replace with an explicit
StateSpace.evaluateTransitions call. Sometimes the evaluation isn't even
needed at all, e. g. if only the transition name is used.