Move non-Groovy-dependent parts into a new prob-java subproject
The de.prob2.kernel project now only contains the parts that are
directly related to Groovy. Everything else is pulled in through a
dependency on prob-java.

This restructuring allows downstream users to use the ProB Java API
without Groovy if they don't need any of the Groovy integration or
scripting parts.