Make TLCModelCheckingOptions independent of StateSpace
All ProB commands are now executed as part of TLCModelChecker, not
beforehand when creating the TLCModelCheckingOptions. This is important
for ProB 2 UI, where the options are often manipulated on the UI thread,
which should not send commands to ProB to avoid hangs. The checker
itself runs on a background thread, so it's okay to send commands there.