Allow creating a StateSpace that doesn't own its underlying animator
If a StateSpace doesn't own its animator, then killing the StateSpace
doesn't kill the animator, but only detaches the StateSpace from the
animator (by removing all listeners and preventing the StateSpace from
using the animator again).

This simplifies the implementation of ReusableAnimator, because the
functionality of its custom InternalAnimator is now supported directly
by StateSpace.