Fix an easy case of trace refinement not killing its animator
Multiple parts of the trace refinement code create StateSpaces without
killing them later, but this is the only one that's easy to fix. The
other cases are more difficult to fix, because they create Transition
objects referencing the new StateSpace and return them to the caller,
so simply killing the StateSpace at the end of the method doesn't work.